[refac] Geocoding, 이미지 데이터 처리 리팩토링
