+++ Project17-C-Map/InteractiveClusteringMap/Network/GeocodingNetwork.swift
@@ -7,77 +7,48 @@
 
 import Foundation
 
-enum NetworkError: Error {
-    case url, author, param, unknown
-    
-    var localizedDescription: String {
-        switch self {
-        case .url:
-            return "잘못된 주소 입니다."
-        case .author:
-            return "잘못된 인증키 입니다.."
-        case .param:
-            return "잘못된 파라미터 입니다."
-        case .unknown:
-            return "알 수 없는 에러입니다."
-        }
-    }
-}
-
 protocol Geocodable: class {
     
-    func address(lat: Double, lng: Double, completion: @escaping ((Result<Data, Error>) -> Void))
+    func address(lat: Double, lng: Double, completion: @escaping ((String) -> Void))
     
 }
 
 final class GeocodingNetwork: Geocodable {
     
-    func address(lat: Double, lng: Double, completion: @escaping ((Result<Data, Error>) -> Void)) {
-        guard let url = URL(string: "\(Geocoding.url)" + "\(EndPoint.query(lat: lat, lng: lng))")
-        else {
-            defer {
-                completion(.failure(NetworkError.url))
-            }
-            return
-        }
+    private let provider: DataProvided
+    
+    init(store: DataProvided) {
+        self.provider = store
+    }
+    
+    func address(lat: Double, lng: Double, completion: @escaping ((String) -> Void)) {
+        let path: String = "\(GeocodingURL.url)" + "\(EndPoint.query(lat: lat, lng: lng))"
         
-        guard  let headerField = Header.get() else {
-            defer {
-                completion(.failure(NetworkError.author))
-            }
-            return
-        }
-        
-        var request = URLRequest(url: url)
-        request.allHTTPHeaderFields = headerField
-        
-        URLSession.shared.dataTask(with: request) { (data, response, error) in
-            if let error = error {
-                defer {
-                    completion(.failure(error))
+        provider.data(path: path, header: Header.get()) { result in
+            switch result {
+            case .success(let data):
+                guard let data = data else { return }
+                
+                let decoder = JSONDecoder()
+                let result = try? decoder.decode(Geocoding.self, from: data)
+                
+                DispatchQueue.main.async {
+                    completion(result?.description ?? "")
                 }
-                return
-            }
-            
-            guard let httpResponse = response as? HTTPURLResponse,
-                  200...299 ~= httpResponse.statusCode else {
-                defer {
-                    completion(.failure(NetworkError.unknown))
+                
+            case .failure(let error):
+                DispatchQueue.main.async {
+                    completion(error.localizedDescription)
                 }
-                return
             }
-            
-            guard let data = data else { return }
-            
-            completion(.success(data))
-        }.resume()
+        }
     }
     
 }
 
 private extension GeocodingNetwork {
     
-    enum Geocoding {
+    enum GeocodingURL {
         static let url = "https://naveropenapi.apigw.ntruss.com/map-reversegeocode/v2/gc"
     }
